Fixed the main.go and added base code for existing .json parsing.
