--- API/jsonStructs.go
@@ -1,43 +1,0 @@
-package main
-
-//Portfolio stores all of the information pulled from the portfolio form. This is stored as global variable.
-type Portfolio struct {
-	Information Information
-	About       About
-	Education   Education
-	Project     Project
-	Status      string
-}
-
-//Information stores all of the relevant info on the person creating the portfolio.
-type Information struct {
-	Name  string `json:"Name"`
-	Title string `json:"Title"`
-	Email string `json:"Email"`
-	Phone string `json:"Phone"`
-}
-
-//About stores the text the user enters in the "About Me" field.
-type About struct {
-	Aboutme string `json:"Aboutme"`
-}
-
-//Education stores the college the user attended and the degree they got, if any.
-type Education struct {
-	College string `json:"College"`
-	Degree  string `json:"Degree"`
-}
-
-//Project stores information on the user's projects that they have completed, if any.
-type Project struct {
-	Name string `json:"Name"`
-	Tech string `json:"Tech"`
-	Desc string `json:"Desc"`
-}
-
-// Global variables for each struct for use in different functions.
-var portfolio = Portfolio{}
-var info = Information{}
-var about = About{}
-var education = Education{}
-var project = Project{}
