+++ API/main.go
@@ -4,15 +4,60 @@ import (
 	"encoding/json"
 	"fmt"
 	"html/template"
+	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
 )
 
+//Portfolio stores all of the information pulled from the portfolio form. This is stored as global variable.
+type Portfolio struct {
+	Information Information
+	About       About
+	Education   Education
+	Project     Project
+	Status      string
+}
+
+//Information stores all of the relevant info on the person creating the portfolio.
+type Information struct {
+	Name  string `json:"Name"`
+	Title string `json:"Title"`
+	Email string `json:"Email"`
+	Phone string `json:"Phone"`
+}
+
+//About stores the text the user enters in the "About Me" field.
+type About struct {
+	Aboutme string `json:"Aboutme"`
+}
+
+//Education stores the college the user attended and the degree they got, if any.
+type Education struct {
+	College string `json:"College"`
+	Degree  string `json:"Degree"`
+}
+
+//Project stores information on the user's projects that they have completed, if any.
+type Project struct {
+	Name string `json:"Name"`
+	Tech string `json:"Tech"`
+	Desc string `json:"Desc"`
+}
+
+// Global variables for each struct for use in different functions.
+var portfolio = Portfolio{}
+var info = Information{}
+var about = About{}
+var education = Education{}
+var project = Project{}
+var jsonfile string
+
 func main() {
 	http.Handle("/", http.FileServer(http.Dir(".")))
 	http.HandleFunc("/portfolioform", portfolioform)
 	http.HandleFunc("/formsubmitted", formsubmitted)
+	http.HandleFunc("/editportfolio", editportfolio)
 	fmt.Println("Open localhost:8080")
 	http.ListenAndServe(":8080", nil)
 }
@@ -82,4 +127,31 @@ func formsubmitted(response http.ResponseWriter, request *http.Request) {
 	fmt.Println(portfolio.Project)
 	fmt.Println(portfolio)
 	temp.Execute(response, portfolio)
+}
+
+//The editportfolio function takes info from an existing .json file and displays it in the form at localhost:8080/editportfolio
+func editportfolio(response http.ResponseWriter, request *http.Request) {
+	//Sets the value of the jsonfile variable.
+	jsonfile = "tony.json"
+	//Opens the .json file.
+	readjson, readerr := os.Open(jsonfile)
+	if readerr != nil {
+		panic(readerr)
+	}
+
+	defer readjson.Close()
+	//Reads the .json file.
+	jsonvalue, _ := ioutil.ReadAll(readjson)
+	//Grabs the info out of the .json file.
+	json.Unmarshal(jsonvalue, &portfolio)
+	//Prints the values of variables in the console to make sure the .json file was read properly.
+	fmt.Println(portfolio)
+	fmt.Println(portfolio.About.Aboutme)
+	//Loads the webpage after the .json file is parsed.
+	temp, err := template.ParseFiles("templates/editportfolio.html")
+	if err != nil {
+		panic(err)
+	}
+
+	temp.Execute(response, nil)
 }
