+++ API/templates/editportfolio.html
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+
+<head>
+    <title>
+        Portfolio Form
+    </title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width-device-width, initial-scale=1">
+</head>
+
+<style>
+    body {
+        font-family: Arial, Helvetica, sans-serif;
+    }
+    input {
+        font-size: 14px;
+    }
+    .bigbox {
+        font-family: Arial, Helvetica, sans-serif;
+        font-size: 14px;
+        height: 200px;
+        width: 600px;
+    }
+</style>
+
+<body>
+    <div>
+        <h1>Edit your portfolio</h1>
+        <form action="/formsubmitted" method="POST">
+            <h2>Information</h2>
+            <div>
+                <label for="fullname">Full name</label>
+                <input type="text" id="fullname" name="fullname" placeholder="Full name" required>
+            </div>
+            <div>
+                <label for="jobtitle">Job title</label>
+                <input type="text" id="jobtitle" name="jobtitle" placeholder="Job title" required>
+            </div>
+            <div>
+                <label for="email">E-mail address</label>
+                <input type="text" id="email" name="email" placeholder="E-mail address" required>
+            </div>
+            <div>
+                <label for="phone">Phone number</label>
+                <input type="text" id="phone" name="phone" placeholder="Phone number" required>
+            </div>
+
+            <h2>About me</h2>
+            <div>
+                <label for="aboutme">About me</label><br>
+                <textarea class="bigbox" type="text" id="aboutme" name="aboutme" placeholder="About me" required></textarea>
+            </div>
+
+            <h2>Education</h2>
+            <div>
+                <label for="college">College</label>
+                <input type="text" id="college" name="college" placeholder="College">
+            </div>
+            <div>
+                <label for="degree">Degree</label>
+                <input type="text" id="degree" name="degree" placeholder="Degree">
+            </div>
+
+            <h2>Projects</h2>
+            <div>
+                <label for="projectname">Project name</label>
+                <input type="text" id="projectname" name="projectname" placeholder="Project name">
+            </div>
+            <div>
+                <label for="techused">Technologies used</label>
+                <input type="text" id="techused" name="techused" placeholder="Technologies used">
+            </div>
+            <div>
+                <label for="projectdesc">Project description</label><br>
+                <textarea class="bigbox" type="text" id="projectdesc" name="projectdesc" placeholder="Project description"></textarea>
+            </div>
+
+            <br>
+
+            <button type="submit">Submit portfolio</button>
+        </form>
+    </div>
+</body>
